Grab window when snapshot button pressed
This fixes a bug in delayed snapshots, where the current window was identified at the point the snapshot was created - not at the point the snapshot was requested.

This meant that it wasn't possible to create a screenshot that showed the full QuPath window while interacting with any other window - since only the 'other window' would end up being captured.